Fix handling of verbose toggle in file editor page
