+++ apps/prairielearn/assets/scripts/lib/jobSequenceResults.ts
@@ -4,6 +4,8 @@ import { decodeData, onDocumentReady } from '@prairielearn/browser-utils';
 
 import type { JobSequenceResultsData } from '../../../src/components/JobSequenceResults.js';
 
+import './verboseToggle.js';
+
 onDocumentReady(() => {
   const socket = io();
 
@@ -50,14 +52,5 @@ onDocumentReady(() => {
         }
       },
     );
-  });
-
-  document.querySelectorAll<HTMLInputElement>('.js-toggle-verbose').forEach((checkbox) => {
-    checkbox.addEventListener('change', () => {
-      const targetOutput = document.getElementById(checkbox.getAttribute('data-target-id') ?? '');
-      if (targetOutput) {
-        targetOutput.style.setProperty('--verbose-display', checkbox.checked ? '' : 'none');
-      }
-    });
   });
 });
